requirements.txt
altgraph==0.17.3
anyio==3.6.2
black==23.3.0
certifi==2023.5.7
click==8.1.3
flet==0.7.4
flet-core==0.7.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
macholib==1.16.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
repath==0.9.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
watchdog==2.3.1
websocket-client==1.5.2
websockets==10.4
